# blockchain, integration, mining, main: Rolling merkle root calculation