--- a/blockchain/merkle.go
+++ b/blockchain/merkle.go
@@ -58,14 +58,13 @@ func nextPowerOfTwo(n int) int {
 // HashMerkleBranches takes two hashes, treated as the left and right tree
 // nodes, and returns the hash of their concatenation.  This is a helper
 // function used to aid in the generation of a merkle tree.
-func HashMerkleBranches(left *chainhash.Hash, right *chainhash.Hash) *chainhash.Hash {
+func HashMerkleBranches(left, right *chainhash.Hash) chainhash.Hash {
 	// Concatenate the left and right nodes.
 	var hash [chainhash.HashSize * 2]byte
 	copy(hash[:chainhash.HashSize], left[:])
 	copy(hash[chainhash.HashSize:], right[:])
 
-	newHash := chainhash.DoubleHashH(hash[:])
-	return &newHash
+	return chainhash.DoubleHashH(hash[:])
 }
 
 // BuildMerkleTreeStore creates a merkle tree from a slice of transactions,
@@ -140,13 +139,13 @@ func BuildMerkleTreeStore(transactions []*btcutil.Tx, witness bool) []*chainhash
 		// hashing the concatenation of the left child with itself.
 		case merkles[i+1] == nil:
 			newHash := HashMerkleBranches(merkles[i], merkles[i])
-			merkles[offset] = newHash
+			merkles[offset] = &newHash
 
 		// The normal case sets the parent node to the double sha256
 		// of the concatentation of the left and right children.
 		default:
 			newHash := HashMerkleBranches(merkles[i], merkles[i+1])
-			merkles[offset] = newHash
+			merkles[offset] = &newHash
 		}
 		offset++
 	}