… mining/mining.go