@Roasbeef 2023-05-25
Concept ACK, great work building on the other optimization with some of the utreexo derived fine tuning!