@Roasbeef 2023-08-10
Completed sync tests on mainnet+testnet as a final sanity check!